[Event Hubs Trigger] Update template version
The focus of these changes is to update the version of the package
referenced for the current generation of the Azure SDK.  The version has
been bumped to 6.0.0, with the containing folder renamed to reflect the
change.
