--- Functions.Templates/Templates/EventHubTrigger-CSharp-5.x/.template.config/template.json
@@ -1,71 +1,0 @@
-{
-  "author": "Microsoft",
-  "classifications": [
-    "Azure Function",
-    "Trigger",
-    "EventHub"
-  ],
-  "name": "EventHubTrigger",
-  "identity": "Azure.Function.CSharp.EventHubTrigger.3.x",
-  "groupIdentity": "Azure.Function.EventHubTrigger",
-  "shortName": "eventhub",
-  "tags": {
-    "language": "C#",
-    "type": "item"
-  },
-  "sourceName": "EventHubTriggerCSharp",
-  "symbols": {
-    "namespace": {
-      "description": "namespace for the generated code",
-      "replaces": "Company.Function",
-      "type": "parameter"
-    },
-    "Connection": {
-      "description": "The name of the app setting containing your Event Hub connection string. The connection string must have Send permissions.",
-      "type": "parameter",
-      "defaultValue": "",
-      "replaces": "ConnectionValue"
-    },
-    "EventHubName": {
-      "description": "This is the name of the event hub to which the event will be sent.",
-      "type": "parameter",
-      "defaultValue": "samples-workitems",
-      "replaces": "eventHubNameValue"
-    }
-  },
-  "primaryOutputs": [
-    {
-      "path": "EventHubTriggerCSharp.cs"
-    }
-  ],
-  "defaultName": "EventHubTriggerCSharp",
-  "postActions": [
-    {
-      "Description": "Adding Reference to Microsoft.Azure.WebJobs.Extensions.EventHubs NuGet package",
-      "ActionId": "B17581D1-C5C9-4489-8F0A-004BE667B814",
-      "ContinueOnError": "true",
-      "ManualInstructions": [],
-      "args": {
-        "referenceType": "package",
-        "reference": "Microsoft.Azure.WebJobs.Extensions.EventHubs", "version": "5.4.0",
-        "projectFileExtensions": ".csproj"
-      }
-    },
-    {
-      "description": "Restore NuGet packages required by this project.",
-      "manualInstructions": [],
-      "actionId": "210D431B-A78B-4D2F-B762-4ED3E3EA9025",
-      "continueOnError": true
-    },
-    {
-      "condition": "(HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-      "description": "Opens the function class file in the editor",
-      "manualInstructions": [],
-      "actionId": "84C0DA21-51C8-4541-9940-6CA19AF04EE6",
-      "args": {
-        "files": "0"
-      },
-      "continueOnError": true
-    }
-  ]
-}
