--- Functions.Templates/Templates/EventHubTrigger-CSharp-5.x/function.json
@@ -1,13 +1,0 @@
-{
-  "bindings": [
-    {
-      "type": "eventHubTrigger",
-      "name": "events",
-      "direction": "in",
-      "eventHubName": "samples-workitems",
-      "cardinality": "many",
-      "connection": "",
-      "consumerGroup": "$Default"
-    }
-  ]
-}
