--- Functions.Templates/Templates/EventHubTrigger-CSharp-5.x/metadata.json
@@ -1,17 +1,0 @@
-{
-  "defaultFunctionName": "EventHubTrigger",
-  "description": "$EventHubTrigger_description",
-  "name": "Azure Event Hub trigger",
-  "language": "C#",
-  "category": [
-    "$temp_category_core",
-    "$temp_category_dataProcessing"
-  ],
-  "categoryStyle": "eventHub",
-  "enabledInTryMode": false,
-  "userPrompt": [
-    "connection",
-    "eventHubName",
-    "consumerGroup"
-  ]
-}
