+++ Functions.Templates/Templates/EventHubTrigger-CSharp-6.x/.template.config/vs-2017.3.host.json
@@ -1,0 +1,38 @@
+{
+  "name": {
+    "text": "Event Hub trigger",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+    "id": "2040"
+  },
+  "description": {
+    "text": "A C# function that will be run whenever an event hub receives a new event",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+    "id": "2041"
+  },
+  "order": 500,
+  "icon": "vs-2017.3/eventhub.png",
+  "learnMoreLink": "https://go.microsoft.com/fwlink/?linkid=848145",
+  "uiFilters": [
+    "functions"
+  ],
+  "symbolInfo": [
+    {
+      "id": "Connection",
+      "name": {
+        "id": "2042",
+        "text": "Connection string setting name",
+        "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+      },
+      "isVisible": true
+    },
+    {
+      "id": "EventHubName",
+      "name": {
+        "id": "2043",
+        "text": "Event Hub name",
+        "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+      },
+      "isVisible": true
+    }
+  ]
+}
