+++ Functions.Templates/Templates/EventHubTrigger-CSharp-6.x/.template.config/vs-2017.3/eventhub-mac-32@2x.png
