+++ Functions.Templates/Templates/EventHubTrigger-CSharp-6.x/function.json
@@ -1,0 +1,13 @@
+{
+  "bindings": [
+    {
+      "type": "eventHubTrigger",
+      "name": "events",
+      "direction": "in",
+      "eventHubName": "samples-workitems",
+      "cardinality": "many",
+      "connection": "",
+      "consumerGroup": "$Default"
+    }
+  ]
+}
