+++ Functions.Templates/Templates/EventHubTrigger-CSharp-6.x/sample.dat
