+++ Functions.Templates/Templates/EventHubTrigger-FSharp/.template.config/template.json
@@ -51,7 +51,7 @@
       "ManualInstructions": [],
       "args": {
         "referenceType": "package",
-        "reference": "Microsoft.Azure.WebJobs.Extensions.EventHubs", "version": "5.4.0",
+        "reference": "Microsoft.Azure.WebJobs.Extensions.EventHubs", "version": "6.0.0",
         "projectFileExtensions": ".fsproj"
       }
     },
